+++ tracer/src/Datadog.Trace/DatabaseMonitoring/DatabaseMonitoringPropagator.cs
@@ -46,15 +46,11 @@ namespace Datadog.Trace.DatabaseMonitoring
                 return false;
             }
 
-            if (!injectStoredProcedure && command.CommandType == CommandType.StoredProcedure)
+            if ((!injectStoredProcedure && command.CommandType == CommandType.StoredProcedure)
+                || (command.CommandType == CommandType.StoredProcedure && integrationId != IntegrationId.SqlClient))
             {
-                // we don't want to inject the comment for stored procedures if it has been disabled explicitly
-                return false;
-            }
-
-            if (command.CommandType == CommandType.StoredProcedure && integrationId != IntegrationId.SqlClient)
-            {
-                // we only support stored procedures for SqlClient
+                // We don't inject into StoredProcedures unless enabled as we change the commands
+                // We don't inject into StoredProcedures unless we are in SqlClient
                 return false;
             }
 
@@ -128,11 +124,11 @@ namespace Datadog.Trace.DatabaseMonitoring
                 *    in SqlCommand.cs we have: rpc.rpcName = this.CommandText; // just get the raw command text
                 * Injecting the comment in the CommandText would result in getting an exception that the StoredProcedure name wasn't found (and would look like an empty string).
                 *
+                * We CAN NOT safely do this though if the Command has any parameters that are not Input (e.g. Return, InputOutput, Output).
+                *
                 * What this function does is:
                 *   - Swap to EXEC statement (which is a text command executing a stored procedure).
-                *   - Build a parameter list for the EXEC statement, so that we can pass all the parameters to it.
-                *   - These are the Input and Output parameters, but they may not be present
-                *   - We don't modify the parameters on the Command itself, this will still be handled fine by SQL Server from testing
+                *   - Build a Input parameter list for the EXEC statement, so that we can pass all the parameters to it.
                 * Some helpful links:
                 * reference RPC for stored procedure: https://github.com/microsoft/referencesource/blob/51cf7850defa8a17d815b4700b67116e3fa283c2/System.Data/System/Data/SqlClient/SqlCommand.cs#L5548
                 * reference text SQL (with params): https://github.com/microsoft/referencesource/blob/51cf7850defa8a17d815b4700b67116e3fa283c2/System.Data/System/Data/SqlClient/SqlCommand.cs#L4488
@@ -156,19 +152,19 @@ namespace Datadog.Trace.DatabaseMonitoring
                     }
                 }
 
-                // Save the original stored procedure name
-                string procName = command.CommandText ?? string.Empty;
+                var procName = command.CommandText ?? string.Empty;
 
                 if (string.IsNullOrEmpty(procName))
                 {
                     return false;
                 }
 
-                string quotedName = string.Empty;
-
+                string? quotedName;
                 try
                 {
-                    quotedName = ParseAndQuoteIdentifier(procName, isUdtTypeName: false);
+                    // dbo.SomeName -> [dbo].[SomeName]
+                    // this uses some underlying SqlClient code that I've copied to parse the identifier
+                    quotedName = VendoredSqlHelpers.ParseAndQuoteIdentifier(procName, isUdtTypeName: false);
                     if (string.IsNullOrEmpty(quotedName))
                     {
                         // if we can't parse the identifier, return false
@@ -356,93 +352,6 @@ namespace Datadog.Trace.DatabaseMonitoring
             BinaryPrimitives.WriteUInt64BigEndian(span.Slice(1 + sizeof(ulong) + sizeof(ulong)), traceId.Lower);
 
             return contextBytes;
-        }
-
-        // parse an string of the form db.schema.name where any of the three components
-        // might have "[" "]" and dots within it.
-        // returns:
-        //   [0] dbname (or null)
-        //   [1] schema (or null)
-        //   [2] name
-        // NOTE: if perf/space implications of Regex is not a problem, we can get rid
-        // of this and use a simple regex to do the parsing
-        // https://github.com/dotnet/SqlClient/blob/414f016540932d339054c61abc5ae838401cdb06/src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SqlParameter.cs#L2433
-        internal static string[] ParseTypeName(string typeName, bool isUdtTypeName)
-        {
-            try
-            {
-                string errorMsg = string.Empty;
-                return MultipartIdentifier.ParseMultipartIdentifier(typeName, "[\"", "]\"", '.', 3, true, errorMsg, true);
-            }
-            catch (Exception)
-            {
-                return []; // return empty array if we can't parse the typeName
-            }
-        }
-
-        // https://github.com/dotnet/SqlClient/blob/414f016540932d339054c61abc5ae838401cdb06/src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/SqlCommand.cs#L6496
-        // Adds quotes to each part of a SQL identifier that may be multi-part, while leaving
-        //  the result as a single composite name.
-        private static string ParseAndQuoteIdentifier(string identifier, bool isUdtTypeName)
-        {
-            string[] strings = ParseTypeName(identifier, isUdtTypeName);
-
-            if (strings.Length == 0)
-            {
-                // if we can't parse the identifier, return it as an empty string, we'll check this and not propagate if so
-                return string.Empty;
-            }
-
-            return QuoteIdentifier(strings);
-        }
-
-        // https://github.com/dotnet/SqlClient/blob/414f016540932d339054c61abc5ae838401cdb06/src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/SqlCommand.cs#L6502
-        private static string QuoteIdentifier(ReadOnlySpan<string> strings)
-        {
-            StringBuilder bld = new StringBuilder();
-
-            // Stitching back together is a little tricky. Assume we want to build a full multi-part name
-            //  with all parts except trimming separators for leading empty names (null or empty strings,
-            //  but not whitespace). Separators in the middle should be added, even if the name part is
-            //  null/empty, to maintain proper location of the parts.
-            for (int i = 0; i < strings.Length; i++)
-            {
-                if (0 < bld.Length)
-                {
-                    bld.Append('.');
-                }
-
-                if (strings[i] != null && 0 != strings[i].Length)
-                {
-                    AppendQuotedString(bld, "[", "]", strings[i]);
-                }
-            }
-
-            return bld.ToString();
-        }
-
-        // https://github.com/dotnet/SqlClient/blob/414f016540932d339054c61abc5ae838401cdb06/src/Microsoft.Data.SqlClient/src/Microsoft/Data/Common/AdapterUtil.cs#L547
-        internal static string AppendQuotedString(StringBuilder buffer, string quotePrefix, string quoteSuffix, string unQuotedString)
-        {
-            if (!string.IsNullOrEmpty(quotePrefix))
-            {
-                buffer.Append(quotePrefix);
-            }
-
-            // Assuming that the suffix is escaped by doubling it. i.e. foo"bar becomes "foo""bar".
-            if (!string.IsNullOrEmpty(quoteSuffix))
-            {
-                int start = buffer.Length;
-                buffer.Append(unQuotedString);
-                buffer.Replace(quoteSuffix, quoteSuffix + quoteSuffix, start, unQuotedString.Length);
-                buffer.Append(quoteSuffix);
-            }
-            else
-            {
-                buffer.Append(unQuotedString);
-            }
-
-            return buffer.ToString();
         }
     }
 }
